+++ backend/src/main/java/org/kmurygin/healthycarbs/mealplan/model/Recipe.java
@@ -38,10 +38,10 @@ public class Recipe {
     @ToString.Exclude
     private List<RecipeIngredient> ingredients = new ArrayList<>();
 
-    private Integer calories;
-    private Integer carbs;
-    private Integer protein;
-    private Integer fat;
+    private Double calories;
+    private Double carbs;
+    private Double protein;
+    private Double fat;
 
     @Enumerated(EnumType.STRING)
     private DietType dietType;
